Change buffer size and offset to u128

+++ src/run/blake2b.rs
@@ -50,7 +50,7 @@ impl Clone for Blake2b {
 pub fn blake2bf_instruction(buf: Buffer) -> Buffer {
     let mut b = [0u8; 213];
     for i in 0..b.len() {
-        b[i] = buf.read_byte(i);
+        b[i] = buf.read_byte(i as u128);
     }
     b[212] = b[212] & 0x1u8;
     match eth_blake2bf_from_bytes(b) {
